requirements.txt
pandas==1.3.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.3
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.62.0
typing-extensions==3.10.0.0
zipp==3.5.0
matplotlib==3.4.3
python-dotenv==0.19.0
sklearn
